♻️ refactor [#58] : payment refactor
