+++ src/main/java/com/sparta/spring_deep/_delivery/domain/payment/controller/PaymentController.java
@@ -1,5 +1,8 @@
-package com.sparta.spring_deep._delivery.domain.payment;
+package com.sparta.spring_deep._delivery.domain.payment.controller;
 
+import com.sparta.spring_deep._delivery.domain.payment.dto.PaymentRequestDto;
+import com.sparta.spring_deep._delivery.domain.payment.dto.PaymentResponseDto;
+import com.sparta.spring_deep._delivery.domain.payment.service.PaymentService;
 import com.sparta.spring_deep._delivery.domain.user.details.UserDetailsImpl;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
