+++ src/main/java/com/sparta/spring_deep/_delivery/domain/payment/dto/PaymentRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.spring_deep._delivery.domain.payment;
+package com.sparta.spring_deep._delivery.domain.payment.dto;
 
 import java.math.BigDecimal;
 import lombok.Getter;
