+++ src/main/java/com/sparta/spring_deep/_delivery/domain/payment/repository/PaymentRepository.java
@@ -1,5 +1,6 @@
-package com.sparta.spring_deep._delivery.domain.payment;
+package com.sparta.spring_deep._delivery.domain.payment.repository;
 
+import com.sparta.spring_deep._delivery.domain.payment.model.Payment;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
